Split db functions
